requirements.txt
requests==2.32.5
moviepy==2.2.1
openai-whisper==20250625
google-api-python-client==2.181.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
pillow==11.3.0